% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_metnordic.R
\name{aggregate_day_metnodric}
\alias{aggregate_day_metnodric}
\title{Convert hourly MetNordic files to daily}
\usage{
aggregate_day_metnodric(
  directory,
  variable,
  day,
  outpath,
  overwrite = TRUE,
  preview = TRUE
)
}
\arguments{
\item{directory}{String: Path to the source files}

\item{day}{String: day to convert (ie. "20150901")}

\item{outpath}{String: path to write file}

\item{preview}{Logical: plot map?}
}
\value{
String: path to written file
}
\description{
This function takes 24 hourly files from the same day and converts them to
an daily format
}
\examples{
# TODO

}
