% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{calc_pres_at_alt}
\alias{calc_pres_at_alt}
\title{Estimate Pressure at Altitude}
\usage{
calc_pres_at_alt(
  h,
  talt,
  pref = 101325,
  alpha = -0.0065,
  href = 0,
  g = 9.80665,
  Rs = 287.058
)
}
\arguments{
\item{h}{altitude above sea level in meters}

\item{talt}{temperature at  altitude (\code{h}) Unit: Kelvin}

\item{pref}{static pressure at reference altitude (\code{href}) Unit: Pa}

\item{alpha}{Temperature gradient (negative lapse rate) Unit: K m^-1}

\item{href}{Reference altitude. Unit: m}

\item{g}{Mean gravitational acceleration at sea level. Unit: m s^-2}

\item{Rs}{Specific gas constant of dry air J (kg*K)^-1}
}
\value{
Static pressure at altitude \code{h} in N m^2
}
\description{
Solves for 'p', Static pressure at altitude 'h' in N m-2 assuming a linear
temperature gradient. Modified from Walter Bislin's \href{http://walter.bislins.ch/bloge/index.asp?page=Barometric+Formula}{write up}.
}
\details{
\deqn{p = p_\mathrm{ref}\, \left( {T_\mathrm{ref} + \alpha \cdot (h - h_\mathrm{ref}) \over T_\mathrm{ref}} \right) ^ {-\beta} = p_\mathrm{ref}\, \left( 1 + {\alpha \cdot (h - h_\mathrm{ref}) \over T_\mathrm{ref}} \right) ^ {-\beta}}

with: \eqn{\beta = {M \cdot g \over R \cdot \alpha} = { g \over R_\mathrm{S} \cdot \alpha}}

where:

\eqn{p}	= Static pressure at altitude h in N m-2

\eqn{p_\mathrm{ref}} = Static pressure at reference altitude href. For href = 0 m (sea level), pref = 101,325 Pa

\eqn{T_\mathrm{ref}} = Temperature at reference altitude href. For href = 0 m (sea level), Tref = 288.15 K (15 °C)

\eqn{\alpha} = Temperature gradient (negative lapse rate) = −0.0065 K -m

\eqn{h} = Altitude above sea level in meter from 0 m up to 11,000 m

\eqn{h_\mathrm{ref}} = Reference altitude. For sea level href = 0 m

\eqn{M} = Molar mass; for dry air = 28.9644 g/mol

\eqn{g} = Mean gravitational acceleration at sea level = 9.80665 m/s2

\eqn{R} = Universal Gas Constant = 8.31446 J/(mol·K)

\eqn{R_\mathrm{S}} = Specific Gas Constant; dry air = 287.058 J/(kg·K)

To calculate \eqn{T_\mathrm{ref}} we need to use the lapse rate to estimate
\code{talt}. This can be done with the following formula:

\deqn{T(h) = T_\mathrm{ref} + \alpha_i\, (h-h_\mathrm{ref})}

where:

\eqn{T(h)} = Temperature at altitude h in Kelvin

\eqn{T_\mathrm{ref}} = Temperature at a reference altitude (e.g. sea level)

\eqn{\alpha_\mathrm{i}} = Temperature Gradient (negative Lapse Rate) of the i-th layer in K/m (Kelvin per meter)

\eqn{h} = Altitude in m

\eqn{h_\mathrm{ref}} = Refrence altitude (sea level = 0 m)
}
\examples{
#TODO
}
\author{
Walter Bislin \href{http://walter.bislins.ch/}{(walter.bislins.ch)}, adapted by Moritz Shore
}
