% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoa_tool.R
\name{cocoa_hru}
\alias{cocoa_hru}
\title{Generates a connectivity plot for an HRU}
\usage{
cocoa_hru(
  hru_id,
  buildR_dir,
  agri_pattern = "a_",
  directory = NULL,
  save_to_png = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{hru_id}{(integer) ID of HRU to plot}

\item{buildR_dir}{(string) path to SWATbuildR project directory (containing the 'data/*' subdirectory)}

\item{agri_pattern}{(string) a pattern which identifies agriculturial fields (REGEX)}

\item{directory}{(string) directory where to save plots (optional, only if \code{save_to_png} is enabled. Will be created to sub-directory 'hru_report/'. If left blank, will use working directory).}

\item{save_to_png}{(logical) \code{TRUE} = saves to file, \code{FALSE} = returns plot}

\item{verbose}{(logical) Print status?}
}
\value{
Either the path to the plot if \code{save_to_png} is \code{TRUE}, or returns the ggplot object itself if \code{save_to_png} is \code{FALSE}
}
\description{
There are some important details which need to be observed for using this function, they are listed below.
}
\details{
\enumerate{
\item This function only works with SWATbuildR output, along side with a generated \code{land_connections_as_lines.shp} file which can be generated using \href{https://github.com/MR-Eini/Mini_setup_CREATE/blob/main/1_Setup/Libraries/create_connectivity_line_shape.R}{this code}.
\item The land use theme is set up for my specific catchment. If you need to add to this theme then please send a pull request with the additional legend entries.
}
}
\seealso{
\code{\link[=cocoa_vis]{cocoa_vis()}}
}
