% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoa_tool.R
\name{cocoa_vis}
\alias{cocoa_vis}
\title{Visualize COCOA on a catchment scale}
\usage{
cocoa_vis(buildR_dir, directory = NULL, RERENDER = TRUE, verbose = TRUE)
}
\arguments{
\item{buildR_dir}{(string) path to SWATbuildR project directory (containing the 'data/*' subdirectory)}

\item{directory}{(string) directory where to save the generated HRU plots}

\item{RERENDER}{(logical) re-generate the HRU plots?}

\item{verbose}{(logical) print status?}
}
\value{
\code{mapview} map of HRUs in the catchment. Click to view each HRUs connectivity.
}
\description{
NOTE: this function has a few requirements to work, please see below.
}
\details{
\enumerate{
\item This function only works with SWATbuildR output, along side with a generated \code{land_connections_as_lines.shp} file which can be generated using \href{https://github.com/MR-Eini/Mini_setup_CREATE/blob/main/1_Setup/Libraries/create_connectivity_line_shape.R}{this code}.
\item The land use theme is set up for my specific catchment. If you need to add to this theme then please send a pull request with the additional legend entries.
}
}
\seealso{
\code{\link[=cocoa_hru]{cocoa_hru()}}
}
