% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnrav3_ncdf.R
\name{convert_to_cwatm}
\alias{convert_to_cwatm}
\title{Convert to CWatM format}
\usage{
convert_to_cwatm(input, output, verbose = FALSE)
}
\arguments{
\item{input}{location of files generated by \code{mnrv3_hourly_to_daily_ncdf4}}

\item{output}{location of files to be generated by this function.}

\item{verbose}{print status?}
}
\value{
the path of generated files
}
\description{
This function takes daily files created by \code{mnrv3_hourly_to_daily_ncdf4} and
converts them into the format required for CWatM
}
