% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5_swatplus.R
\name{era5_swatplus}
\alias{era5_swatplus}
\title{Convert ERA5 data from open meteo to SWAT+ input.}
\usage{
era5_swatplus(
  open_meteo_path,
  extract_path,
  swat_setup,
  aux_data = NULL,
  epsg_code = NULL,
  verbose = TRUE,
  selected_vars = c("max_temp", "min_temp", "precipitation_amount", "rad", "windspeed",
    "rh"),
  write_wgn = TRUE,
  clean_files = TRUE,
  sqlite_path = NULL,
  fill_missing = FALSE
)
}
\arguments{
\item{open_meteo_path}{path to file downloaded from open-meteo in daily format with the correct variables}

\item{extract_path}{path to location where data can be extracted and processed}

\item{swat_setup}{path to swat+ setup}

\item{aux_data}{if auxilliary data is to be added, path to xlsx file using \code{SWATprepR} formatting.}

\item{epsg_code}{See \code{SWATprepR} docs}

\item{verbose}{Print? default is \code{TRUE}}

\item{selected_vars}{Variables to incorporate from ERA5. default is all of them}

\item{write_wgn}{See \code{SWATprepR}docs}

\item{clean_files}{See \code{SWATprepR}docs}

\item{sqlite_path}{See \code{SWATprepR}docs}

\item{fill_missing}{See \code{SWATprepR} docs}
}
\value{
Nothing
}
\description{
\code{open_meteo_path} needs to point to a file downloaded from open-meteo.
}
