% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnorenal3.R
\name{get_metno_reanalysis3}
\alias{get_metno_reanalysis3}
\title{Download data from the metno renanalysis project}
\usage{
get_metno_reanalysis3(
  area,
  directory = NULL,
  fromdate = "2012-09-01 10:00:00",
  todate = "2012-09-01 20:00:00",
  area_buffer = 1500,
  grid_resolution = NULL,
  preview = TRUE
)
}
\arguments{
\item{area}{(string) path to geo-referenced shapefile covering the desired area}

\item{directory}{(string) path to desired working directory (default: working directory)}

\item{fromdate}{(string) date and time for start of time series (ie. "2012-09-01 10:00:00")}

\item{todate}{(string) date and time for end of time series (ie. "2013-09-01 10:00:00")}

\item{area_buffer}{desired buffer around the provided shapefile (in meters, default 1500)}

\item{grid_resolution}{(integer) desired resolution of downloaded grid in kilometers. (see help page for more details)}

\item{preview}{generate graphs showing previews of data download? (boolean)}
}
\value{
Writes .csv files into a folder located at the provided directory.
One .csv file for each grid point within the (buffered) shape file area.
Addtionally one metadata file (.csv) is written with the attributes of each
other file
}
\description{
This function accesses the THREDDS server of met.no with download queries
formatted from user input. (to be expanded upon)
}
\details{
The \code{grid_resolution} parameter lets you choose how many stations to select
from the grid. If for example you pass a "3", then every 3 station will be
selected in both the horizontal and vertical axis, giving you a grid with
stations placed 3km apart from each other on a regular grid. This can help
with downloading data from huge catchments where 1x1km grid size is not
necessary. Please note, no averaging of the "non-selected" grid cells is
performed.
}
\examples{
 # for demonstration purposes, use path of package
 ##example_file_path <- system.file(package = "miljotools", "/extdata/metno_reanal/watershed.shp")

 ##get_metno_reanalysis3(
 ##area = example_file_path,
 ##fromdate = "2015-01-01",
 ##todate = "2015-01-02",
 ##area_buffer = 100,
 ##preview = TRUE
 ##)


}
\author{
Moritz Shore
}
