% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_aggregate.R
\name{metnordic_aggregate}
\alias{metnordic_aggregate}
\title{Convert hourly MetNordic files to daily}
\usage{
metnordic_aggregate(
  directory,
  variable,
  method,
  day,
  outpath,
  overwrite = TRUE,
  preview = TRUE
)
}
\arguments{
\item{directory}{String: Path to the source files}

\item{variable}{String: variable to aggregate}

\item{method}{method of aggregation (mean, min, max, sum)}

\item{day}{String: day to convert (ie. "20150901")}

\item{outpath}{String: path to directory of to be created files}

\item{overwrite}{Logical: overwrite existing file?}

\item{preview}{Logical: plot map?}
}
\value{
String: path to written file
}
\description{
This function takes 24 hourly files from the same day and variable and
converts them to a daily format
}
\examples{
# TODO

}
