% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_aggregate_daterange.R
\name{metnordic_aggregate_daterange}
\alias{metnordic_aggregate_daterange}
\title{Aggregate a date range}
\usage{
metnordic_aggregate_daterange(
  directory,
  variable,
  method,
  start,
  end,
  outpath,
  overwrite = TRUE,
  verify = FALSE,
  n_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{directory}{String: Path to the source files (as downloaded by \code{metnordic_download()} or \code{metnordic_download_daterange()})}

\item{variable}{String: MET Nordic variable to aggregate (eg. "precipitation_amount", \href{https://github.com/metno/NWPdocs/wiki/MET-Nordic-dataset#parameters}{(see more)})}

\item{method}{String: method of aggregation ("mean", "min", "max", "sum")}

\item{start}{String: start of the date range to aggregate (eg. "2015-01-01")}

\item{end}{String: end of the date range to aggregate (eg. "2015-12-31")}

\item{outpath}{String: path to directory of to be created files}

\item{overwrite}{Logical: overwrite existing files? (optional, default TRUE)}

\item{n_cores}{Numeric: max number of cores to perform operation with. (optional)}

\item{verbose}{Logical: print to console? (optional)}

\item{verfiy}{Logical: verify the existence of all required files? (recommended to be TRUE)}
}
\value{
Returns a named list with file paths to each aggregated file. \code{FALSE}
indicates a file failed to be aggregated, likely due to not having 24hrs
(24 individual files) to aggregated from.
}
\description{
A wrapper function for \code{metnordic_aggregate()} which aggregates files from a
start to end date (in parallel). You must supply a variable type and a method
of aggregation (eg. mean). This function is designed to take input from
\code{metnordic_download_daterange()} and provides input for
\code{metnordic_merge_daily()}
}
\details{
Currently supported are the following types of aggregation: "mean" "min"
"max" "sum". Operations are performed in paralell on multiple cores. You can
control the number of cores used with the \code{num_cores} parameters.
}
\seealso{
\code{\link[=metnordic_aggregate]{metnordic_aggregate()}} \code{\link[=metnordic_download_daterange]{metnordic_download_daterange()}} \code{\link[=metnordic_merge_daily]{metnordic_merge_daily()}}
}
