% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_download_daterange.R
\name{metnordic_download_daterange}
\alias{metnordic_download_daterange}
\title{Download MET Nordic files}
\usage{
metnordic_download_daterange(queries, directory, mn_variables, verbose = FALSE)
}
\arguments{
\item{queries}{list as passed by \code{metnordic_buildqueries()}}

\item{directory}{folder where to download}

\item{mn_variables}{MET Nordic variables to download.}

\item{verbose}{print status?}
}
\value{
path to download directory
}
\description{
Downloads a list of provided queries from the MET No thredds server using the
OPENDAP protocol. The queries should be passed from
\code{metnordic_buildqueries()}. Folder of where to download, and which variables
to download also need to be provided! Any files already present in the folder
will not be re-downloaded. This means that if the download fails for whatever
network reason, you can just restart the function and it should pick off
where it left off. If you are having issues with downloads, make sure to
check https://status.met.no/ for server (THREDDS) status.
}
\seealso{
\code{\link[=metnordic_buildquery]{metnordic_buildquery()}}
}
