% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_merge.R
\name{metnordic_merge}
\alias{metnordic_merge}
\title{Merge MetNordic Files}
\usage{
metnordic_merge(folderpath, variable, outpath, overwrite = FALSE)
}
\arguments{
\item{folderpath}{outpath of files from \code{metnordic_aggregate()}}

\item{variable}{MetNordic variable to be used}

\item{outpath}{Folderpath where to write files}

\item{overwrite}{overwrite existing file?}
}
\value{
file path of generated files
}
\description{
This function merges daily MetNordic files into a single file for a single
variable. The input data should be the output data of \code{metnordic_aggregate()}
}
\examples{
# TODO
}
