% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_merge_hourly.R
\name{metnordic_merge_hourly}
\alias{metnordic_merge_hourly}
\title{Merge MET Nordic files (hourly)}
\usage{
metnordic_merge_hourly(folderpath, variable, outpath, overwrite = FALSE)
}
\arguments{
\item{folderpath}{(String) folder where individual files are located}

\item{variable}{(String) MET Nordic variable to combine}

\item{outpath}{(String) folder where to write the file}

\item{overwrite}{(Boolean) overwrite existing file?}
}
\value{
path of written file
}
\description{
This function lets you combine the hourly files downloaded by
\code{download_metnordic()} into single files. \strong{Note}, these files are separate
per variable!
}
\seealso{
\code{\link[=metnordic_download]{metnordic_download()}}
}
