% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_csv.R
\name{metnordic_point}
\alias{metnordic_point}
\title{Download MET Nordic data for one point}
\usage{
metnordic_point(
  area,
  path = NULL,
  fromdate = "2012-09-01 10:00:00",
  todate = "2012-09-01 20:00:00",
  mn_variables = NULL,
  dataset = "continuous",
  verbose = TRUE
)
}
\arguments{
\item{area}{(string) path to geo-referenced shapefile (point) of the desired area. (optionally, you can pass a \code{sf} object directly.)}

\item{path}{(string) path to download directory}

\item{fromdate}{(string) date and time for start of time series (ie. "2012-09-01 10:00:00")}

\item{todate}{(string) date and time for end of time series (ie. "2013-09-01 10:00:00")}

\item{mn_variables}{(vector) Leave blank for default variables}

\item{dataset}{(string) from which dataset to source the files from? ("reanalysis", "operational", or "continuous" for both).}

\item{verbose}{(boolean) generate graphs showing previews of data download?}
}
\value{
Function returns a path to where .csv files of the download were
written. One .csv file for each grid point within the (buffered) shape
file area. Additionally one metadata file (.csv) is written with the
}
\description{
note: this function only works for point geometry in csv format.
}
\details{
This is a modified version of the legacy \code{get_metno_reanalysis3} designed
specifically for downloading single points in csv format.
}
\examples{

 if(FALSE){
 metnordic_csv(
 area = example_file_path,
 fromdate = "2015-01-01",
 todate = "2015-01-02",
 verbose = TRUE
 )
 }


}
\author{
Moritz Shore
}
