% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_reproject.R
\name{metnordic_reproject}
\alias{metnordic_reproject}
\title{Reproject MetNordic Data}
\usage{
metnordic_reproject(filepath, outfile, projstring = NULL)
}
\arguments{
\item{filepath}{path to .nc file to be reprojected}

\item{outfile}{filepath of .nc file to be created. (if the same file path is used, .nc file will be overwritten)}

\item{projstring}{desired projection in \href{https://epsg.io/docs}{proj4 format}. By default, \href{https://epsg.io/32633}{UTM33N} will be used.}
}
\value{
filepath to reprojected file
}
\description{
Takes in MetNordic .nc files as created by \code{metnordic_download()} and
reprojects them to desired projection.
}
\examples{
# TODO
}
