% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_verify_download.R
\name{metnordic_verify_download}
\alias{metnordic_verify_download}
\title{Verify MET Nordic download}
\usage{
metnordic_verify_download(folder)
}
\arguments{
\item{folder}{}
}
\value{
Nothing
}
\description{
This function scans the download folder for files that are "too small" and
then deletes them. You can then use the \code{download_metnordic_daterange()} with
the same directory (and queries from \code{metnordic_build_query()}) function to
re-download the deleted ones
}
\examples{

# TODO
}
