% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mt_print}
\alias{mt_print}
\title{custom print function for miljotools}
\usage{
mt_print(verbose, function_name, text, text2 = NULL, rflag = FALSE)
}
\arguments{
\item{verbose}{print or not}

\item{function_name}{function name string}

\item{text}{string 1}

\item{text2}{string 2}

\item{rflag}{in place replacement flag}
}
\value{
nothing
}
\description{
custom print function for miljotools
}
\keyword{internal}
