% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnrav3_ncdf.R
\name{read_write_ncdf}
\alias{read_write_ncdf}
\title{Read and write ncdf files}
\usage{
read_write_ncdf(url, savefiles, directory, foldername, verbose = FALSE)
}
\arguments{
\item{url}{urls to download}

\item{savefiles}{paths to save}
}
\value{
path to downloaded files
}
\description{
instead of coverting them to dataframe format
}
\keyword{internal}
