% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnorenal3.R
\name{reanalysis3_swatinput}
\alias{reanalysis3_swatinput}
\title{Create SWAT+ meteo input from MetNo Reanalysis3 data}
\usage{
reanalysis3_swatinput(path, sqlite_path, outpath = NULL, verbose = FALSE)
}
\arguments{
\item{path}{path to daily data provided by \code{reanalysis3_daily()}}

\item{sqlite_path}{path to your SWAT+ sqlite file.}

\item{outpath}{optional path to directory where .xlsx file will be written.}

\item{verbose}{print status?}
}
\value{
path to generated files.
}
\description{
Takes data gathered by \code{get_metno_reanalysis3()} and downscaled by
\code{reanalysis3_daily()} and creates SWAT+ meteo input files and weather
generators.
}
