% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seNorge.R
\name{senorge_buildquery}
\alias{senorge_buildquery}
\title{Build SeNorge2018 Nordic Download Query}
\usage{
senorge_buildquery(
  bounding_coords,
  variables,
  fromdate,
  todate,
  grid_resolution = 1,
  verbose = FALSE
)
}
\arguments{
\item{bounding_coords}{list as returned by \code{metnordic_coordwindow()}}

\item{variables}{character vector of variables to download may include the following: "tn", "tx", "rr", "tg"}

\item{fromdate}{date from when to start downloading data from (eg. "1957-01-01")}

\item{todate}{date from when to end downloading data from (eg. "2024-12-31")}

\item{grid_resolution}{integer value fo grid resolution (eg 1 for a 1x1km grid)}

\item{verbose}{logical, print?}
}
\value{
Returns a named list with the urls and filenames which is to be passed to \code{senorge_download()}.
}
\description{
This function builds the URL queries for downloading SeNorge data through
the OPENDAP protocol. The requirements for this function to work are the
bounding coordinates as divined by function \code{metnordic_coordwindow()}, the
variables of interest, the starting and ending dates and the desired grid
resolution. The results of this function can be downloaded when passed to
\code{senorge_download()}.
}
\examples{
# TODO
}
\seealso{
\code{\link[=metnordic_coordwindow]{metnordic_coordwindow()}} \code{\link[=senorge_download]{senorge_download()}}
}
