% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seNorge.R
\name{senorge_download}
\alias{senorge_download}
\title{Download SeNorge2018 data}
\usage{
senorge_download(
  queries,
  directory = NULL,
  variables = NULL,
  polygon = NULL,
  verbose = FALSE
)
}
\arguments{
\item{queries}{(list) generated by metnordic_buildqueries()}

\item{directory}{(char) folder where to download (will be created if it does not exist)}

\item{variables}{(char, vector) variables to download ("tn", "tx", "rr", "tg")}

\item{polygon}{(char) path to same polygon shapefile as passed to \code{metnordic_coordwindow()}}

\item{verbose}{(flag) print status?}
}
\value{
path to downloaded files
}
\description{
This function downloads the files as queried by \code{metnordic_buildquery()}. If
you do not pass a \code{directory}, files will be downloaded in the working
directory. You have the option of downloading the following SeNorge2018
variables at daily resolution:
\itemize{
\item \code{tn} (minimum daily temperature in Celsius)
\item \code{tx} (maximum daily temperature in Celsius)
\item \code{rr} (sum daily precipitation in mm),
\item \code{tg} (mean daily temperature in Celsius)
If you do not provide a vector of variables to download, all will be downloaded.
You can also provide a \code{polygon} (Georeferenced simple feature, the same that
you provided to \code{metnordic_coordwindow()}) to verify that the download is
covering your desired area.
}
}
