% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_polygons.R
\name{simplify_polygons}
\alias{simplify_polygons}
\title{Simplify Polygons}
\usage{
simplify_polygons(polygon_map, type, interactive = FALSE, verbose = TRUE)
}
\arguments{
\item{polygon_map}{shapefile containing a column \code{flag} of type \code{string} with labels for every polygon as either 'flagged' or 'ok'. Flagged polygons will be simplified.}

\item{type}{character string of the name of the \code{type} column in your shapefile. Polygons of the same type will be prioritized for merging If you do not need this feature, simply set the type column to your ID column.}

\item{interactive}{\code{TRUE/FALSE} Setting this parameter to true will let you inspect the changes made.}

\item{verbose}{print actions to console?}
}
\value{
Returns a modified version of \code{polygon_map} with the simplifications.
}
\description{
This function merges flagged polygons into neighboring polygons with the
longest shared border. It will prioritize neighboring polygons of the same
type. See parameter descriptions for more details.
}
\examples{

# See Vignette
}
