% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_metnordic.R
\name{swap_metnordic}
\alias{swap_metnordic}
\title{METNordic input for SWAP}
\usage{
swap_metnordic(dldir, outpath, timescale = "daily", verbose = FALSE)
}
\arguments{
\item{dldir}{download directory containing data and metadata, as generated by \code{metnordic_csv()}}

\item{outpath}{desired location for the swap .met file to be written}

\item{timescale}{"daily" or "hourly". (hourly not supported yet)}

\item{verbose}{print some stuff to console?}
}
\value{
path to written .met file
}
\description{
This function converts Meteo data from MET Nordic into a SWAP-compatible
format. Note: need to use \code{metnordic_csv()} to create proper input for this
function.
}
\seealso{
\code{\link[=metnordic_csv]{metnordic_csv()}}
}
