% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnorenal3.R
\name{swat_weather_input_chain}
\alias{swat_weather_input_chain}
\title{Generate SWAT+ weather input for any watershed in the nordics.}
\usage{
swat_weather_input_chain(
  area,
  swat_setup,
  directory = NULL,
  from = NULL,
  to = NULL,
  area_buffer = NULL,
  verbose = TRUE,
  precision = 2,
  write_wgn = TRUE,
  sqlite_path = NULL
)
}
\arguments{
\item{area}{The catchment area to retrieve data for. (must be a shapefile)}

\item{swat_setup}{The path to your SWAT+ setup (input files, aka TxtInOut)}

\item{directory}{directory to download and process data in}

\item{from}{start of the to-be-dowloaded timeseries (ie. and min: "2012-09-01
10:00:00")}

\item{to}{end of the to-be-dowloaded timeseries (ie. and max: "2023-01-31
10:00:00")}

\item{area_buffer}{optional buffer in meters around the provided area}

\item{verbose}{print status messages?}

\item{precision}{Optional, which precision (integer) should the hourly data
be rounded down to when converted from daily to hourly. Default is '2'
decimal places.}

\item{write_wgn}{would you like to calculate and write the weather generator?}

\item{sqlite_path}{optionally, you can pass the path of your .sqlite file in
order to update the database with your new met files}
}
\description{
This function combines 3 \code{miljotools} functions in a single gridded data
retrieval and processing pipeline to write and assign weather data to a SWAT+
setup. The functions involved are
}
\details{
\enumerate{
\item \code{get_metno_reanalysis3()} downloads and processes the hourly gridded
reanalysis data for the nordics
\item \code{reanalysis3_daily()} converts these hourly timeseries into daily.
\item \code{reanalysis3_swatinput()} converts these timeseries into a SWAT+
compatible format as well as generating the weather generator, and updating
SWAT+ input files with the help of the R-package \code{SWARTprepR}
}

For more details please see the help pages of the individual functions. Also
please note the the package \code{SWATprepR} is required for this pipeline.
}
\author{
Moritz Shore, Svajunas Plunge
}
