% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_metnordic.R
\name{swatplus_metnordic}
\alias{swatplus_metnordic}
\title{Apply MetNordic meteo data to a SWAT+ setup}
\usage{
swatplus_metnordic(
  directory,
  swat_setup,
  write_wgn = TRUE,
  start = NA,
  end = NA,
  sqlite_path = NULL,
  backup = FALSE,
  aux_data = NULL,
  epsg_code = NULL,
  fill_missing = TRUE,
  clean_files = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{directory}{path to weather data as created by \code{metnordic_extract_grid()}}

\item{swat_setup}{path to SWAT+ setup}

\item{write_wgn}{should the weather generator be written?}

\item{start}{optional parameter to define start date of time series}

\item{end}{optional parameter to define end date of time series4}

\item{sqlite_path}{path to your SWAT+ sqlite file (only needed if you wish to update your database). Warning: start and end parameters will be ignored in this case (SWATprepR limitation)}

\item{backup}{(logical, defaults to true) creates a backup of your swat folder before modification}

\item{aux_data}{Additional (non-metnordic) data to add to your SWAT+ project. Must be a path to an .xlsx file which matches the \code{SWATprepR} format of \code{load_template()}.}

\item{epsg_code}{\code{SWATprepR}: (optional) Integer, EPSG code for station coordinates. Default epsg_code = 4326, which stands for WGS 84 coordinate system.}

\item{fill_missing}{\code{SWATprepR}: (optional) Boolean, TRUE - fill data for missing stations with data from closest stations with available data. FALSE - leave stations without data. Weather generator will be used to fill missing variables for a model. Default fill_missing = TRUE.}

\item{clean_files}{\code{SWATprepR}: Logical, if \code{TRUE}, will remove all existing weather files in model setup folder before writing new ones. Default \code{clean_files = TRUE}.}

\item{verbose}{print to console?}
}
\value{
path to SWAT+ setup
}
\description{
This function takes the output from \code{metnordic_extract_grid()} and applies it
to a SWAT+ setup, with the help of \href{https://biopsichas.github.io/SWATprepR/}{SWATprepR}.
}
\seealso{
\code{\link[=metnordic_extract_grid]{metnordic_extract_grid()}}
}
