% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermopluviograms.R
\name{thermopluviogram}
\alias{thermopluviogram}
\title{Thermopluviogram}
\usage{
thermopluviogram(
  modelled_climate,
  observed_pr,
  observed_tm,
  outpath,
  location,
  ref_startdate,
  ref_enddate,
  obs_startdate,
  obs_enddate,
  verbose = TRUE,
  chosen_model_runs = NULL,
  experiments = c("hist", "rcp45", "rcp85"),
  left_yaxis_only = FALSE,
  fixed_axis = TRUE,
  axis_font_size = 7,
  yaxisrange = c(800, 1100),
  xaxisrange = c(4.5, 8.5),
  xlab = "Mean Annual Temperature (°C)"
)
}
\arguments{
\item{modelled_climate}{filepath of climate files (i.e. "Climate_Data/Modelled/")}

\item{observed_pr}{filepath of observed precipitation data (.csv)}

\item{observed_tm}{filepath of the observed mean temperature data (.csv)}

\item{outpath}{outpath for plotting}

\item{location}{location of plot (is used for file naming) (string)}

\item{ref_startdate}{reference period start (ie."1971-01-01")}

\item{ref_enddate}{reference period d end (ie. "2005-12-31")}

\item{obs_startdate}{observation period start (ie. "2041-01-01")}

\item{obs_enddate}{observation period end (ie. "2070-12-31")}

\item{verbose}{print progress?}

\item{chosen_model_runs}{Names of the models to color as blue}

\item{experiments}{Experiments to generate (ie. c("hist", "rcp45","rcp85"))}

\item{left_yaxis_only}{If set to true, only the historical plot axis will be labeled (This is so the three plots can be placed next to each other in a figure)}

\item{fixed_axis}{use a fixed axis range when plotting? (boolean)}

\item{axis_font_size}{Font size (Integer)}

\item{yaxisrange}{y axis range (ie. c(800,1100), mm)}

\item{xaxisrange}{x axis range (ie.  c(4.5, 8.5), Celsius)}

\item{xlab}{x axis label. (ie. "Mean Annual Temperature)", u00B0 is degree symbol)}
}
\value{
path to generated files
}
\description{
Combines climate data from \href{https://klimaservicesenter.no/}{Norsk Klima Service Center} and observed data to produce
\href{https://www.dwd.de/EN/ourservices/jahresthermo/jahresthermo.html}{"Thermopluviograms"}.
}
\details{
https://www.senorge.no/PrecTempMap

todo: describe file struture

Generates thermopluviograms and extreme rain index plots from NVE
not flexible. only works with current source data structure.

Date:  01.06.2022
Author:  Moritz Shore
}
