% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwatm_convert_nc.R
\name{cwatm_convert_nc}
\alias{cwatm_convert_nc}
\title{Convert MetNordic Files to CWatM Meteo Input}
\usage{
cwatm_convert_nc(infile, outfile)
}
\arguments{
\item{infile}{path to projected MetNordic .nc file as created by \code{metnordic_reproject()}}

\item{outfile}{path and file name to desired output file}
}
\value{
path to outfile
}
\description{
This function converts a projected metnodric file to be compatible with
CWatm. This function mainly just changes the dimension names from easting and
northing to x and y but has a crap load of boiler plate because... well..
ncdf4.
}
\examples{
#TODO
}
