% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_aggregate.R
\name{metnordic_aggregate}
\alias{metnordic_aggregate}
\title{Convert hourly MET Nordic files to daily}
\usage{
metnordic_aggregate(
  directory,
  variable,
  method,
  day,
  outpath,
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{directory}{String: Path to the source files (as downloaded by \code{metnordic_download()} or \code{metnordic_download_daterange()})}

\item{variable}{String: MET Nordic variable to aggregate (ie. "precipitation_amount", \href{https://github.com/metno/NWPdocs/wiki/MET-Nordic-dataset#parameters}{(see more)})}

\item{method}{String: method of aggregation ("mean", "min", "max", "sum")}

\item{day}{String: day to convert ("YYYYMMDD" format, ie. "20150901")}

\item{outpath}{String: path to directory of to be created files}

\item{overwrite}{Logical: overwrite existing file?}

\item{verbose}{Logical: plot map?}
}
\value{
String: path to written file
}
\description{
This function takes 24 hourly files from the same day and variable and
converts them to a daily format. You have the choice of four different
aggregation methods: min, max, mean, and sum.
}
\details{
This function is designed to take input from \code{metnordic_download()} and
provides input for \code{metnordic_merge()}

Code largely adapted from this handy guide:
(\href{https://pjbartlein.github.io/REarthSysSci/netCDF.html#create-and-write-a-projected-netcdf-file}{link})
from Pat Bartlein, bartlein@uoregon.edu
}
\examples{
# TODO

}
\seealso{
\code{\link[=metnordic_merge_daily]{metnordic_merge_daily()}} \code{\link[=metnordic_download]{metnordic_download()}} \code{\link[=metnordic_download_daterange]{metnordic_download_daterange()}}
}
\author{
Moritz Shore
}
