% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_buildquery.R
\name{metnordic_buildquery}
\alias{metnordic_buildquery}
\title{Build MetNordic Download Query}
\usage{
metnordic_buildquery(
  bounding_coords,
  mn_variables,
  fromdate,
  todate,
  grid_resolution,
  verbose
)
}
\arguments{
\item{bounding_coords}{as determined by \code{metnordic_coordwindow()}}

\item{mn_variables}{MetNordic variables (see \href{https://github.com/metno/NWPdocs/wiki/MET-Nordic-dataset#parameters}{documentation})}

\item{fromdate}{ie. "2019-01-01 00:00:00"}

\item{todate}{ie. "2020-12-31 23:00:00"}

\item{grid_resolution}{an integer, ie. 3 for 3x3 km grid.}

\item{verbose}{print to console?}
}
\value{
charater vector of all the OPENDAP URLs to download.
}
\description{
This function builds the URL queries for downloading MetNordic Data through
the OPENDAP protocol. The requirements for this function to work are the
bounding coordinates as divined by function \code{metnordic_coordwindow()}, the
variables of interest, the starting and ending dates and the desired grid
resolution
}
\examples{
# TODO
}
