% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_coordwindow.R
\name{metnordic_coordwindow}
\alias{metnordic_coordwindow}
\title{Get MetNordic Coordinate Window}
\usage{
metnordic_coordwindow(area_path, area_buffer, preview)
}
\arguments{
\item{area_path}{path to shapefile of extent / point (this must have point or polygon geometry!)}

\item{area_buffer}{buffer in m to place around extent / point}

\item{preview}{Logical: preview the coord window?}
}
\value{
returns a list of the min and max x and y cells for downloading.
}
\description{
This function retrieves the coordinate window for downloading MetNordic files
}
\examples{
# TODO
}
