% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_coordwindow.R
\name{metnordic_coordwindow}
\alias{metnordic_coordwindow}
\title{Get a MET Nordic Coordinate Window}
\usage{
metnordic_coordwindow(area_path, area_buffer = 0, verbose = FALSE)
}
\arguments{
\item{area_path}{String: path to shapefile of region / point (this must have point or polygon geometry!)}

\item{area_buffer}{Integer: buffer in m to place around shapefile / point}

\item{verbose}{Logical: plot the coordinate window?}
}
\value{
returns a list of the min and max x and y cells for downloading.
}
\description{
This function retrieves the coordinate window from a shapefile for
downloading MET Nordic files. This window can then be passed to
\code{metnordic_buildquery()} to parse OPENDAP urls to download.
}
\examples{
# TODO
}
\seealso{
\code{\link[=metnordic_buildquery]{metnordic_buildquery()}}
}
\author{
Moritz Shore
}
