% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_metnordic.R
\name{metnordic_download}
\alias{metnordic_download}
\title{Downloads MetNordic Data}
\usage{
metnordic_download(url, outdir, vars, overwrite = FALSE, preview = TRUE)
}
\arguments{
\item{url}{String: OPDAP url which the function should download}

\item{outdir}{String: Location where the .nc file should be written.}

\item{vars}{Vector: Variables to extract. All Must be present in the URL}

\item{overwrite}{Logical: Overwrite files if they already exist?}

\item{preview}{Logical: Print preview?}
}
\value{
String: filepaths to generated files
}
\description{
Downloads Metnordic reanlysis data when passed a URL and desired variables.
}
\details{
Following this handy guide: \href{https://pjbartlein.github.io/REarthSysSci/netCDF.html#create-and-write-a-projected-netcdf-file}{link}
}
\examples{
#TODO
}
