% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_extract.R
\name{metnordic_extract}
\alias{metnordic_extract}
\title{MET Nordic Extract - Point Timeseries regional downloads}
\usage{
metnordic_extract(
  directory,
  mn_variables,
  point,
  outdir,
  name,
  verbose = FALSE
)
}
\arguments{
\item{directory}{(String) directory containing merged files as created by \code{metnordic_merge_hourly()}}

\item{mn_variables}{(vector, strings) variables to extract}

\item{point}{an object of class "sf" with point geometry.}

\item{outdir}{(string) directory in which to write the file (.csv)}

\item{name}{(string) name of the file (will be added to filename)}

\item{verbose}{(boolean) print?}
}
\value{
path to written file
}
\description{
This function extracts timeseries from (hourly) files from the chain
\code{metnordic_coordwindow()} -->  \code{metnordic_buildquery()} -->
\code{metnordic_coordmetnordic_download()} --> \code{metnordic_merge_hourly()} -->
\code{metnordic_extract()}. This function extracts from the nearest grid cell, if
you would like a bi-linear interpolation of the nearest 4 cells, please try
\code{metnordic_extract_grid()}
}
