% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_extract_grid.R
\name{metnordic_extract_grid}
\alias{metnordic_extract_grid}
\title{Extract data on a (ir)regular grid basis}
\usage{
metnordic_extract_grid(
  merged_path,
  area,
  buffer = 0,
  mn_variables,
  outdir,
  verbose
)
}
\arguments{
\item{merged_path}{String, Path to merged ncdf files from \code{metnordic_merge_hourly()}}

\item{area}{Geo-referenced shapefile either of polygon or point geometry. Passing a polygon geometry will lead to a regular grid being extracted from all overlaying points.}

\item{buffer}{Numeric, buffer in meters. Useful for getting grid cells just outside of catchment. Not used if a point geometry is passed.}

\item{mn_variables}{Character Vector, Met Nordic variables to extract}

\item{outdir}{String, Folder where data will be written}

\item{verbose}{Boolean, print status}
}
\value{
nothing
}
\description{
This function takes the files from \code{metnordic_merge_hourly()} and extracts
timeseries at desired locations. If a polygon shapefile is passed to the
function, data will be extracted at each grid cells within the (buffered)
area. If a point geometry is supplied, data will be extracted at each point. A
file is written to disk with .csv file for each grid cell and a metadata file for each grid cell.
}
\details{
TODO: add resolution to polygon grid..
}
