% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_extract_grid.R
\name{metnordic_extract_grid}
\alias{metnordic_extract_grid}
\title{extract grid of metnordic (WIP)}
\usage{
metnordic_extract_grid(
  merged_path,
  area,
  mn_variables,
  outdir,
  verbose,
  regular = TRUE
)
}
\arguments{
\item{merged_path}{path to merged ncdf files}

\item{area}{shape file, either an area to be gridded, or a premade grid (regular = FALSE!)}

\item{mn_variables}{variables to extract}

\item{outdir}{where to extract}

\item{verbose}{you know it}

\item{regular}{create a grid? or one was provided? (to be replaced with geometry detection)}
}
\value{
nothing
}
\description{
extract grid of metnordic (WIP)
}
