% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnordic_merge_hourly.R
\name{metnordic_merge_hourly}
\alias{metnordic_merge_hourly}
\title{Merge MET Nordic files (hourly)}
\usage{
metnordic_merge_hourly(
  folderpath,
  variable,
  outpath,
  n_cores = NULL,
  overwrite = FALSE,
  verify = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{folderpath}{(String) folder where individual files are located}

\item{variable}{(String) MET Nordic variable to combine}

\item{outpath}{(String) folder where to write the file}

\item{n_cores}{(Integer) Number of cores to use for parallel processing (Defaults to max - 2)}

\item{overwrite}{(Boolean) overwrite existing file?}

\item{verify}{(Boolean) an optional check to see if all files to be merged are incremental (Recommended!)}

\item{verbose}{(Boolean) print status?}
}
\value{
path of written file
}
\description{
This function lets you combine the hourly files downloaded by
\code{download_metnordic()} into single files. \strong{Note}, these files are separate
per variable!
}
\seealso{
\code{\link[=metnordic_download]{metnordic_download()}} \code{\link[=metnordic_extract]{metnordic_extract()}}
}
