% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnorenal3.R
\name{reanalysis3_daily}
\alias{reanalysis3_daily}
\title{Downscale MetNo Reanalysis3 data to daily resolution (Legacy)}
\usage{
reanalysis3_daily(path, outpath = NULL, verbose = FALSE, precision = 2)
}
\arguments{
\item{path}{path to output of get_metno_reanalysis3()}

\item{outpath}{(optional) path to write the new files to}

\item{verbose}{(optional) flag to print status}

\item{precision}{round the values to what precision? (integrer, default 2)}
}
\value{
path of written files
}
\description{
\strong{NOTE:} This is a legacy function and is no longer supported.
}
\details{
This function takes the hourly data from get_metno_reanalysis3() and
recalculates it into daily (temporal) resolution.
}
