% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_metnordic.R
\name{reanalysis3_swatinput}
\alias{reanalysis3_swatinput}
\title{Create SWAT+ meteo input from MetNo Reanalysis3 data (SUPERSEDED)}
\usage{
reanalysis3_swatinput(
  path,
  swat_setup,
  write_wgn = TRUE,
  start = NA,
  end = NA,
  sqlite_path = NULL,
  verbose = FALSE,
  backup = TRUE
)
}
\arguments{
\item{path}{path to daily data provided by \code{reanalysis3_daily()}}

\item{swat_setup}{path to your SWAT+ setup. (Required!)}

\item{write_wgn}{calculate and write the weather generator? defaults to true. (for now just based on station #1 (bottom left))}

\item{start}{optional parameter to define start date of time series}

\item{end}{optional parameter to define end date of time series}

\item{sqlite_path}{path to your SWAT+ sqlite file (only needed if you wish to
update your database). Warning: start and end parameters will be ignored in this case (SWATprepR limitation)}

\item{verbose}{print status?}

\item{backup}{(logical, defaults to true) creates a backup of your swat folder before modification}
}
\value{
Files are generated in provided paths
}
\description{
\strong{SUPERSEDED}. Note: intended use of this function is within \code{swatplus_metnordic()}
}
\seealso{
\code{\link[=swatplus_metnordic]{swatplus_metnordic()}}

Takes data gathered by \code{get_metno_reanalysis3()} and downscaled by
\code{reanalysis3_daily()} and creates SWAT+ meteo input files and weather
generators for your SWAT+ input with help from SWATprepR.

\strong{NOTE} Default parameter value for \code{mn_variables} in \code{get_metno_reanalysis3()}  is required for this to work.

\strong{NOTE} package \code{SWATprepR} is required for this function

https://github.com/biopsichas/SWATprepR
}
\author{
Moritz Shore, Svajunas Plunge
}
