% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seNorge.R
\name{senorge_extract_grid}
\alias{senorge_extract_grid}
\title{Extract data from SeNorge2018 files}
\usage{
senorge_extract_grid(directory, outdir, area, variables, buffer, verbose)
}
\arguments{
\item{directory}{WIP}

\item{outdir}{WIP}

\item{area}{WIP}

\item{variables}{WIP}

\item{buffer}{WIP}

\item{verbose}{WIP}
}
\value{
WIP
}
\description{
WIP
}
\examples{

WIP
}
