% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatplus_metnordic.R
\name{swatplus_metnordic}
\alias{swatplus_metnordic}
\title{Apply MetNordic meteo data to a SWAT+ setup}
\usage{
swatplus_metnordic(
  directory,
  swat_setup,
  write_wgn = TRUE,
  start = NA,
  end = NA,
  sqlite_path = NULL,
  verbose
)
}
\arguments{
\item{directory}{path to weather data as created by \code{metnordic_extract_grid()}}

\item{swat_setup}{path to SWAT+ setup}

\item{write_wgn}{should the weather generator be written?}

\item{start}{optional parameter to define start date of time series}

\item{end}{optional parameter to define end date of time series4}

\item{sqlite_path}{path to your SWAT+ sqlite file (only needed if you wish to update your database). Warning: start and end parameters will be ignored in this case (SWATprepR limitation)}

\item{verbose}{}
}
\value{
path to SWAT+ setup
}
\description{
This function takes the output from \code{metnordic_extract_grid()} and applies it
to a SWAT+ setup, with the help of \href{https://biopsichas.github.io/SWATprepR/}{SWATprepR}.
}
\seealso{
\code{\link[=metnordic_extract_grid]{metnordic_extract_grid()}}
}
